removed dt parameter from some updates

+++ MoonRuntime/Source/Window/WindowSpecification.hpp
@@ -5,7 +5,7 @@
 struct WindowSpecification
 {
     std::string Title = "Title";
-    float Width = 1280;
-    float Height = 720;
+    int Width = 1280;
+    int Height = 720;
     bool VSync = true;
 };
